New Oscillator class
